2021 - day 2

+++ 2021/day02/test.h
@@ -4,5 +4,5 @@
 
 namespace aoc2021_day02 {
     int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    long long part_2(std::string_view path);
 }
